[533249] Minimize code regions with "@generated NOT"
The classes BasicDecorationNodeImpl and DecorationNodeImpl completely
hand-written. These classes and their subclasses are missing the data
fields for a number of features, so the generator generates wrong code
for the eIsSetGen() methods.

I addressed the compile errors with a trick. To all of the five
offending eIsSetGen() methods I added "@generated NOT_CDO". To the
EMF generator that is exactly the same as "@generated NOT", but later
I can simply have the "NOT_CDO" removed by an Ant script and the
generator will generate correct code for CDO
(with reflective feature delegation).

Change-Id: I45ae10a82388319207f0f37dafe77794a0cfc8e5
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=533249
Signed-off-by: Eike Stepper <stepper@esc-net.de>